[Template-Engine]: Add test shouldThrowExceptionIfPlaceholderValueIsMissing
