--- src/main/java/com/example/messenger/TemplateEngine.java
@@ -1,15 +1,0 @@
-package com.example.messenger;
-
-import java.util.Map;
-
-public class TemplateEngine {
-
-    public String render(String template, Map<String, String> values) {
-        String result = template;
-        for (Map.Entry<String, String> entry : values.entrySet()) {
-            String placeholder = "#{" + entry.getKey() + "}";
-            result = result.replace(placeholder, entry.getValue());
-        }
-        return result;
-    }
-}
