+++ src/main/java/com/example/messenger/exception/MissingPlaceholderValueException.java
@@ -1,0 +1,7 @@
+package com.example.messenger.exception;
+
+public class MissingPlaceholderValueException extends RuntimeException {
+    public MissingPlaceholderValueException(String placeholder) {
+        super("Missing value for placeholder: " + placeholder);
+    }
+}
