+++ src/main/java/com/example/messenger/service/TemplateEngine.java
@@ -1,0 +1,27 @@
+package com.example.messenger.service;
+
+import com.example.messenger.exception.MissingPlaceholderValueException;
+
+import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class TemplateEngine {
+
+    private static final Pattern PLACEHOLDER_PATTERN = Pattern.compile("#\\{(.+?)\\}");
+
+    public String render(String template, Map<String, String> values) {
+        Matcher matcher = PLACEHOLDER_PATTERN.matcher(template);
+        StringBuilder result = new StringBuilder();
+        while (matcher.find()) {
+            String key = matcher.group(1);
+            if (!values.containsKey(key)) {
+                throw new MissingPlaceholderValueException(key);
+            }
+            String replacement = values.get(key);
+            matcher.appendReplacement(result, Matcher.quoteReplacement(replacement));
+        }
+        matcher.appendTail(result);
+        return result.toString();
+    }
+}
